+++ src/ports/repositories/account-repository.ts
@@ -1,0 +1,6 @@
+import type { Account } from "../../core/accounts/account.js";
+
+export type AccountRepository = {
+  list(): Promise<Account[]>;
+  save(account: Account): Promise<void>;
+};